Move message functions into new messages module.
